+++ src/MeshBlock.h
@@ -495,7 +495,8 @@ public:
     void outputOrphan(FILE* fp, int i)
     {
         fprintf(
-            fp, "%f %f %f\n", rxyz[3 * i], rxyz[3 * i + 1], rxyz[3 * i + 2]);
+            fp, "%f %f %f\n", rxyz[static_cast<int>(3 * i)], rxyz[3 * i + 1],
+            rxyz[3 * i + 2]);
     }
     void clearOrphans(HOLEMAP* holemap, int nmesh, const int* itmp);
     void clearOrphans(ADAPTIVE_HOLEMAP* holemap, int nmesh, int* itmp);
